Move related functions to transform.js
