+++ src/fontra/client/core/transform.js
@@ -176,6 +176,67 @@ function _unpackTransformObject(t) {
   return [t.xx, t.xy, t.yx, t.yy, t.dx, t.dy];
 }
 
+export function decomposedToTransform(transformation) {
+  let t = new Transform();
+  t = t.translate(
+    transformation.translateX + transformation.tCenterX,
+    transformation.translateY + transformation.tCenterY
+  );
+  t = t.rotate(transformation.rotation * (Math.PI / 180));
+  t = t.scale(transformation.scaleX, transformation.scaleY);
+  t = t.skew(
+    transformation.skewX * (Math.PI / 180),
+    transformation.skewY * (Math.PI / 180)
+  );
+  t = t.translate(-transformation.tCenterX, -transformation.tCenterY);
+  return t;
+}
+
+export function decomposedFromTransform(affine) {
+  // Decompose a 2x2 transformation matrix into components:
+  // - rotation
+  // - scaleX
+  // - scaleY
+  // - skewX
+  // - skewY
+  const [a, b, c, d] = [affine.xx, affine.xy, affine.yx, affine.yy];
+  const delta = a * d - b * c;
+
+  let rotation = 0;
+  let scaleX = 0,
+    scaleY = 0;
+  let skewX = 0,
+    skewY = 0;
+
+  // Apply the QR-like decomposition.
+  if (a != 0 || b != 0) {
+    const r = Math.sqrt(a * a + b * b);
+    rotation = b > 0 ? Math.acos(a / r) : -Math.acos(a / r);
+    [scaleX, scaleY] = [r, delta / r];
+    [skewX, skewY] = [Math.atan((a * c + b * d) / (r * r)), 0];
+  } else if (c != 0 || d != 0) {
+    const s = Math.sqrt(c * c + d * d);
+    rotation = Math.PI / 2 - (d > 0 ? Math.acos(-c / s) : -Math.acos(c / s));
+    [scaleX, scaleY] = [delta / s, s];
+    [skewX, skewY] = [0, Math.atan((a * c + b * d) / (s * s))];
+  } else {
+    // a = b = c = d = 0
+  }
+
+  const transformation = {
+    translateX: affine.dx,
+    translateY: affine.dy,
+    rotation: rotation * (180 / Math.PI),
+    scaleX: scaleX,
+    scaleY: scaleY,
+    skewX: skewX * (180 / Math.PI),
+    skewY: skewY * (180 / Math.PI),
+    tCenterX: 0,
+    tCenterY: 0,
+  };
+  return transformation;
+}
+
 const _EPSILON = 1e-15;
 const _ONE_EPSILON = 1 - _EPSILON;
 const _MINUS_ONE_EPSILON = -1 + _EPSILON;
